requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0
